requirements.txt
Flask~=3.1.0
python-dotenv~=1.0.1
pymongo~=4.10.1
sentence_transformers
google-generativeai==0.6.0
IPython~=8.29.0
flask-cors~=5.0.0
pydantic==2.7.4
openai==1.35.3
vertexai==1.49.0
google-cloud-aiplatform==1.49.0
mistralai~=0.4.2
fastembed==0.3.1
protobuf~=4.25.5
sentence-transformers~=3.3.0
numpy~=1.26.4